requirements.txt
mysqlclient~=1.4.4
Flask~=2.0.1
twilio~=7.2.0

python-dotenv~=0.19.1